# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Adpoint API",
        "contact": {
            "email": "alkhamilnaz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        },
        {
            "url": "https://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a listing of the products",
                "operationId": "list products",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/product/add": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add Product",
                "operationId": "Add Product",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Display a listing of the product by id",
                "operationId": "list product by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Registrasi"
                ],
                "summary": "Register user",
                "operationId": "Registration user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "ktp": {
                                        "type": "string"
                                    },
                                    "npwp": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductSchema": {
                "title": "Products",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "Billboard Cilandak"
                    },
                    "added_by": {
                        "type": "string",
                        "example": "seller"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "20"
                    },
                    "category_id": {
                        "type": "string",
                        "example": "5"
                    },
                    "subcategory_id": {
                        "type": "string",
                        "example": "2"
                    },
                    "brand_id": {
                        "type": "string",
                        "example": "7"
                    },
                    "photos": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "uploads/foto1.jpg"
                        },
                        "collectionFormat": "multi",
                        "enum": [
                            "uploads\\/products\\/photos\\/img1.png",
                            "uploads\\/products\\/photos\\/img2.png"
                        ]
                    },
                    "thumbnail_img": {
                        "type": "string",
                        "example": "uploads/thumbnail_img.png"
                    },
                    "featured_img": {
                        "type": "string",
                        "example": "uploads/featured_img.png"
                    },
                    "flash_deal_img": {
                        "type": "string",
                        "example": "uploads/flash_deal_img.png"
                    },
                    "tags": {
                        "type": "string",
                        "example": "billboard,baleho,cilandak"
                    },
                    "description": {
                        "type": "text",
                        "example": "<p>description</p>"
                    },
                    "video_provider": {
                        "type": "string",
                        "example": "youtobe"
                    },
                    "video_link": {
                        "type": "string",
                        "example": "https://www.youtube.com/watch?v=qn8HDW68cIk&t=979s"
                    },
                    "unit_price": {
                        "type": "string",
                        "example": "30000000"
                    },
                    "choice_options": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Periode"
                                },
                                "options": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Bulanan"
                                    },
                                    "collectionFormat": "multi",
                                    "enum": [
                                        "Bulanan",
                                        "Tahunan"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    },
                    "variations": {
                        "properties": {
                            "Harian": {
                                "properties": {
                                    "price": {
                                        "type": "string",
                                        "example": "1000"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "BH-Harian"
                                    },
                                    "qty": {
                                        "type": "string",
                                        "example": "1000"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "tax": {
                        "type": "string",
                        "example": "10"
                    },
                    "tax_type": {
                        "type": "string",
                        "example": "amount"
                    },
                    "discount": {
                        "type": "string",
                        "example": null
                    },
                    "discount_type": {
                        "type": "string",
                        "example": null
                    },
                    "meta_title": {
                        "type": "string",
                        "example": "Billboard Cilandak"
                    },
                    "meta_description": {
                        "type": "string",
                        "example": "Billboard Cilandak"
                    },
                    "meta_img": {
                        "type": "string",
                        "example": "uploads/meta_img.png"
                    },
                    "pdf": {
                        "type": "string",
                        "example": "uploads/document.pdf"
                    },
                    "slug": {
                        "type": "string",
                        "example": "Billboard-Cilandak"
                    },
                    "latlong": {
                        "type": "string",
                        "example": "-6.881291,107.58345299999996"
                    },
                    "alamat": {
                        "type": "text",
                        "example": "Komplek Setra Sari Mall Blok C3 No. 35"
                    },
                    "provinsi": {
                        "type": "string",
                        "example": "Jawa Barat"
                    },
                    "kota": {
                        "type": "string",
                        "example": "Bogor"
                    },
                    "kecamatan": {
                        "type": "string",
                        "example": "Leuwiliang"
                    },
                    "audien_target": {
                        "type": "text",
                        "example": "pengguna jalan"
                    },
                    "statistik_masyarakat": {
                        "type": "text",
                        "example": "Wirausaha:orang|Swasta:orang"
                    },
                    "jumlah_pendengarradio": {
                        "type": "string",
                        "example": "20"
                    },
                    "target_pendengarradio": {
                        "type": "string",
                        "example": "Wirausaha:orang|Swasta:orang"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}